backend/requirements.txt
python-dotenv
azure-cosmos
openai
fastapi
uvicorn
python-multipart
jinja2
starlette
langchain
langchain-openai
azure-identity
gunicorn
flask-limiter
azure-search-documents==11.4.0
azure-communication-email
applicationinsights